The program now successfully parses input from std::cin.
The mistake in the previous commit wasn't due to the std::strtok(...) function
or my implementation of it; I completely forgot I had to use std::getline
instead of just reading directly with the '<<' operator to std::cin. Once I
fixed that, the program was successfully reading the entire input regardless of
spaces, and the std::strtok function -which was working fine- did exactly what
was expected.